test(theme): update tests
